auto documentation - restructuring the project by creating isogeo_search_engine directory
