--- api/__init__.py
@@ -1,1 +1,0 @@
-from . import api_client
