+++ isogeo_search_engine/__init__.py
@@ -1,0 +1,3 @@
+from .main_search_engine import IsogeoSearchEngine
+from .api import IsogeoAPI
+from .ui import Interface
