+++ isogeo_search_engine/api/__init__.py
@@ -1,0 +1,1 @@
+from .api_client import IsogeoAPI
