+++ isogeo_search_engine/ui/__init__.py
@@ -1,0 +1,1 @@
+from .ui_objs import Interface
